Adding acquisition grouping logic
