+++ docker/Dockerfile.dev
@@ -9,6 +9,7 @@ ARG CVUTILS_REPO_DIR=$REPO_DIR/cvutils
 RUN mkdir -p $LAB_DIR $REPO_DIR $DATA_DIR
 
 RUN apt-get update && apt-get install -y --no-install-recommends git vim
+RUN pip install --upgrade pip
 
 # Install these specific package versions to avoid their annoying breaking changes
 RUN pip install keras==2.1.6
